refactor: consolidate logger usage
