+++ __tests__/hooks/chat/useQuickCommands.test.tsx
@@ -31,7 +31,7 @@ const mockQuickCommands = [
 ];
 
 // ロガーのモック
-jest.mock('@/utils/logger', () => ({
+jest.mock('@/utils/common', () => ({
   logger: {
     info: jest.fn(),
   },
