+++ __tests__/unit/tools/instrument-type-tools.test.ts
@@ -3,12 +3,12 @@
 // Cascade: Corrected import path for ExchangeType.
 
 import { changeInstrumentTypeTool } from '../../../src/mastra/tools/instrument-type-tools';
-import { logger } from '../../../utils/logger';
+import { logger } from '../../../utils/common';
 import fetch from 'node-fetch'; // Import to mock
 import { ExchangeType, ProductType } from '@/types/api';
 
 // Mock logger
-jest.mock('../../../utils/logger', () => ({
+jest.mock('../../../utils/common', () => ({
   logger: {
     info: jest.fn(),
     error: jest.fn(),
