+++ services/data/chart-data-service.ts
@@ -11,7 +11,7 @@ import { BitgetApiClient } from '../api/bitget/client.new';
 import { IChartDataService } from './interfaces';
 import { OHLCData, Timeframe } from '../../types/chart';
 import { ExchangeType } from '@/types/api';
-import { logger } from '../../utils/logger';
+import { logger } from '../../utils/common';
 import { normalizeSymbol } from '../../utils/formatters';
 import { getSocketService } from '../socket/index';
 import { cacheService } from '../cache/service';
