+++ services/socket/socket-service.ts
@@ -11,7 +11,7 @@ import { OrderBookData } from '../../types/market';
 import { OHLCData, Timeframe } from '../../types/chart';
 import { ProductType } from '@/types/api';
 import { BitgetApiClient } from '../api/bitget/client.new';
-import { logger } from '../../utils/logger';
+import { logger } from '../../utils/common';
 import { 
   ISocketService, 
   IWebSocketClient,
