+++ services/socket/subscription-manager.ts
@@ -11,7 +11,7 @@ import { OrderBookData } from '../../types/market';
 import { OHLCData, Timeframe } from '../../types/chart';
 import { ProductType } from '@/types/api';
 import { normalizeSymbol } from '../../lib/utils';
-import { logger } from '../../utils/logger';
+import { logger } from '../../utils/common';
 import { ISubscriptionManager, IWebSocketClient } from './interfaces';
 import { getWebSocketClient } from './websocket-client';
