--- utils/logger.ts
@@ -1,8 +1,0 @@
-/**
- * @deprecated このファイルは utils/common/logger.ts に移動しました。
- * 新しい場所から import してください。
- * 例: import { logger } from '@/utils/common';
- */
-
-export { logger } from './common';
-export { default } from './common/logger';
